Try relocating TokenStateTranslator
