+++ hedera-node/hedera-mono-service/src/test/java/com/hedera/node/app/service/mono/state/merkle/MerkleTokenSerdeTest.java
@@ -16,7 +16,6 @@
 
 package com.hedera.node.app.service.mono.state.merkle;
 
-import com.hedera.node.app.service.mono.state.migration.TokenStateTranslator;
 import com.hedera.test.serde.EqualityType;
 import com.hedera.test.serde.SelfSerializableDataTest;
 import com.hedera.test.utils.SeededPropertySource;
