+++ hedera-node/hedera-mono-service/src/test/java/com/hedera/node/app/service/mono/utils/UtilsConstructorTest.java
@@ -51,6 +51,7 @@ import com.hedera.node.app.service.mono.sigs.utils.ImmutableKeyUtils;
 import com.hedera.node.app.service.mono.sigs.utils.MiscCryptoUtils;
 import com.hedera.node.app.service.mono.sigs.utils.PrecheckUtils;
 import com.hedera.node.app.service.mono.state.merkle.MerkleAccount;
+import com.hedera.node.app.service.mono.state.merkle.TokenStateTranslator;
 import com.hedera.node.app.service.mono.state.merkle.internals.BitPackUtils;
 import com.hedera.node.app.service.mono.state.merkle.internals.ByteUtils;
 import com.hedera.node.app.service.mono.state.migration.MapMigrationToDisk;
@@ -99,6 +100,7 @@ import org.junit.jupiter.api.Test;
 class UtilsConstructorTest {
     private static final Set<Class<?>> toBeTested = new HashSet<>(Arrays.asList(
             EntityIoUtils.class,
+            TokenStateTranslator.class,
             OrderedComparison.class,
             RecordParsers.class,
             Units.class,
