+++ hedera-node/hedera-mono-service/src/test/java/com/hedera/test/serde/EqualityType.java
@@ -16,7 +16,21 @@
 
 package com.hedera.test.serde;
 
+/**
+ * Used by the base {@link SelfSerializableDataTest} to signal to expected object providers what type of
+ * equality test will be used.
+ *
+ * <p>We added this so that when {@code OBJECT_EQUALITY} is being used, we can insert an extra step of
+ * converting the expected object to-and-from the corresponding PBJ type <I>before</I> returning it.
+ * (This is an efficient way to confirm our bidirectional converters are not changing object semantics.)
+ */
 public enum EqualityType {
+    /**
+     * A test is using object equality to compare expected and actual objects.
+     */
     OBJECT_EQUALITY,
+    /**
+     * A test is using serialized bytes equality to compare expected and actual objects.
+     */
     SERIALIZED_EQUALITY
 }
